+++ CHANGELOG.md
@@ -2,9 +2,33 @@
 
 All notable changes to this project will be documented in this file.
 
-The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-_1.0.0_
-- **[FEATURES]**
-  - ...
+## [1.0.0] - 2024-07-09
+
+### **Added**
+  - Device actions
+    - lock
+    - unlock
+    - isLocked
+  - Keyboard actions
+    - hideKeyboard
+    - isKeyboardShown
+  - Gesture actions
+    - scroll
+    - swipe
+  - Tap actions
+    - tap
+    - doubleTap
+    - longTap
+  - Session actions
+    - terminateApp
+    - activateApp
+    - isAppInstalled
+    - removeApp
+    - installApp
+    - clearApp
+    - backgroundApp
+    - acceptAlert
+    - dismissAlert